Move lifting of replication transform to transform dir
